requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.7
colorama==0.4.6
constantly==23.10.4
cricapi==1.2
cryptography==46.0.7
daphne==4.2.1
distro==1.9.0
Django==4.2.30
django-cors-headers==4.9.0
exceptiongroup==1.3.1
groq==1.2.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kagglehub==1.0.0
kagglesdk==0.1.19
lxml==6.1.0
numpy==2.2.6
packaging==26.1
pandas==2.3.3
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
tomli==2.4.1
tqdm==4.67.3
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
zope.interface==8.3
